Create README.md
This project is a Python-based software solution designed to trace the flow of cryptocurrency transactions through blockchain explorers, with the goal of identifying the probable end user or recipient of a transaction trail. It can be used by cybersecurity professionals, researchers, and investigative agencies to track suspicious or anonymous crypto flows.